requirements.txt
SenseHat
RPi.GPIO
sounddevice
mpu6050